rag-api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pymongo>=4.5.0
google-generativeai>=0.7.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
numpy>=1.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
